requirements.txt
python-telegram-bot==13.15
python-binance==1.0.17
requests